Cleaned up unused code warnings
