+++ src-tauri/src/commands/mod.rs
@@ -1,0 +1,20 @@
+pub mod general;
+mod helpers;
+pub mod rip;
+
+#[macro_export]
+macro_rules! all_commands {
+    () => {
+        tauri::generate_handler!(
+            crate::commands::general::index,
+            crate::commands::general::movie,
+            crate::commands::general::open_browser,
+            crate::commands::rip::rip_one,
+            crate::commands::general::search,
+            crate::commands::general::season,
+            crate::commands::general::the_movie_db,
+            crate::commands::general::tv,
+            crate::commands::rip::rip_season,
+        )
+    };
+}
